Bug fix (check for object type when deleting files for python installs)

+++ models/qc_time_estimator/qc_time_estimator/processing/data_management.py
@@ -93,7 +93,7 @@ def remove_old_pipelines(*, files_to_keep: List[str]):
 
     do_not_delete = files_to_keep + ['__init__.py']
     for model_file in config.TRAINED_MODEL_DIR.iterdir():
-        if model_file.name not in do_not_delete:
+        if model_file.name not in do_not_delete and model_file.is_file():
             model_file.unlink()
